extract byline into standalone component
