+++ packages/frontend/web/components/ArticleBody.tsx
@@ -2,7 +2,6 @@
 
 import React from 'react';
 import { css, cx } from 'emotion';
-import TwitterIcon from '@guardian/pasteup/icons/twitter.svg';
 import { palette } from '@guardian/pasteup/palette';
 import ClockIcon from '@guardian/pasteup/icons/clock.svg';
 import {
@@ -21,6 +20,7 @@ import { ShareCount } from './ShareCount';
 import { SharingIcons } from './ShareIcons';
 import { SubMetaLinksList } from './SubMetaLinksList';
 import { Dateline } from './Dateline';
+import { Byline } from './Byline';
 
 import { MainMedia } from './MainMedia';
 
@@ -342,24 +342,6 @@ const linkColour = pillarMap(
     `,
 );
 
-const profile = css`
-    ${headline(2)};
-    font-weight: 700;
-    margin-bottom: 4px;
-`;
-
-const byline = css`
-    font-style: italic;
-`;
-
-const bylineLink = css`
-    font-style: normal;
-    text-decoration: none;
-    :hover {
-        text-decoration: underline;
-    }
-`;
-
 const ageWarning = css`
     ${textSans(1)};
     display: inline-block;
@@ -368,27 +350,6 @@ const ageWarning = css`
 
     ${leftCol} {
         margin-top: 6px;
-    }
-`;
-
-const twitterHandle = css`
-    ${textSans(1)};
-    font-weight: bold;
-    color: ${palette.neutral[46]};
-
-    padding-right: 10px;
-    display: inline-block;
-
-    svg {
-        height: 10px;
-        max-width: 12px;
-        margin-right: 0px;
-        fill: ${palette.neutral[46]};
-    }
-
-    a {
-        color: ${palette.neutral[46]};
-        text-decoration: none;
     }
 `;
 
@@ -440,59 +401,6 @@ const subMetaSharingIcons = css`
     }
 `;
 
-// this crazy function aims to split bylines such as
-// 'Harry Potter in Hogwarts' to ['Harry Potter', 'in Hogwarts']
-const bylineAsTokens = (bylineText: string, tags: TagType[]): string[] => {
-    const contributorTags = tags
-        .filter(t => t.type === 'Contributor')
-        .map(c => c.title);
-    const regex = new RegExp(`(${contributorTags.join('|')})`);
-
-    return bylineText.split(regex);
-};
-
-const RenderByline: React.FC<{
-    bylineText: string;
-    contributorTags: TagType[];
-    pillar: Pillar;
-}> = ({ bylineText, contributorTags, pillar }) => {
-    const renderedTokens = bylineAsTokens(bylineText, contributorTags).map(
-        (token, i) => {
-            const associatedTags = contributorTags.filter(
-                t => t.title === token,
-            );
-            if (associatedTags.length > 0) {
-                return (
-                    <BylineContributor
-                        contributor={token}
-                        contributorTagId={associatedTags[0].id}
-                        pillar={pillar}
-                        key={i}
-                    />
-                );
-            }
-            return token;
-        },
-    );
-
-    return <>{renderedTokens}</>;
-};
-
-const BylineContributor: React.FC<{
-    contributor: string;
-    contributorTagId: string;
-    pillar: Pillar;
-}> = ({ contributor, contributorTagId, pillar }) => (
-    <a
-        rel="author"
-        className={cx(section, pillarColours[pillar], bylineLink)}
-        data-link-name="auto tag link"
-        href={`//www.theguardian.com/${contributorTagId}`}
-    >
-        {contributor}
-    </a>
-);
-
 export const ArticleBody: React.FC<{
     CAPI: CAPIType;
     config: ConfigType;
@@ -529,27 +437,11 @@ export const ArticleBody: React.FC<{
                     />
                 </div>
                 <div className={cx(meta, guardianLines)}>
-                    <div className={cx(profile, pillarColours[CAPI.pillar])}>
-                        <span className={byline}>
-                            <RenderByline
-                                bylineText={CAPI.author.byline}
-                                contributorTags={CAPI.tags}
-                                pillar={CAPI.pillar}
-                            />
-                        </span>
-                    </div>
-                    {CAPI.author.twitterHandle && (
-                        <div className={twitterHandle}>
-                            <TwitterIcon />
-                            <a
-                                href={`https://www.twitter.com/${
-                                    CAPI.author.twitterHandle
-                                }`}
-                            >
-                                @{CAPI.author.twitterHandle}
-                            </a>
-                        </div>
-                    )}
+                    <Byline
+                        author={CAPI.author}
+                        tags={CAPI.tags}
+                        pillar={CAPI.pillar}
+                    />
                     <Dateline
                         dateDisplay={CAPI.webPublicationDateDisplay}
                         descriptionText={'Published on'}
